Log unique text
Added LogUniqueOriginalText(...) which will send the text to the log file if it is a newly discovered piece of text.
Some pieces of text have rich text tags or weird spaces which make it difficult to determine what to type in for the find string. Now they can be searched for in an external text editor and copied into the config file.