Comment out logUniqueText
Disabled the code that prints each piece of newly detected text to the log. It is no longer in the running for a potential feature, or at least, copy pasting find strings will need to be reworked.
Combined with CrashReporter, this code may upload people's login credentials for BeatLeader and other sensitive info to the cloud for anyone to see. Very, very, extremely very not good.