Revert "Identify text based off of orginal text"
Reverted manually because I had made a few other changes in that commit.
Identify text should not base identification off of original text, it should base off of the current new state. Whatever I saw in the logs for me to originally add this must be dealt with in some other way.