Add config order
Added configOrder to mod config, as well as the functions loadConfigOrder and saveConfigOrder.
The order of configs and entries determine how they are processed. The order of entries are stored in the JSON array, however file loading is not guaranteed to be in the same order every time. Plus, the desired order has to be stored somewhere. As it's user specific, it goes in mod config.