Protoype new UI with createConfigSetting
Start writing new UI by generating each list element with a function.
This is what I did before, but I'm starting to think it may not be the best approach. I want to evaluate the configs when the AnyTextViewController back button is pressed rather than upon immidiate updating of any UI, which means more state and UI complexity. It'd be nice if each config setting had a component, and HMUI::TableView and TableCell is starting to look more appealing the more I look at it.