Add lots of debug logging
Nearly every UI function now logs something... Name and any relevant info.
I was trying to track down a weird crash. This amount of logging is definitely a tad verbose, but I may or may not keep it. If a crash is uploaded, it is useful to see what the person may have done that led up to potentially cause the crash.