Use TextManager::Awake and OnDestroy
Switch from OnEnable and OnDisable to MonoBehaviour functions that more closely align with whether or not the script is attached to the text object.
This fixes a crash that occured when a table cell was disabled and reused. I am still unsure what the *exact* trigger was for `boost:regex_search` to crash, all the pointers checked out, attempting mutex didn't change anything, but it still must've been related to the fact that the new Entry and the EntryTableCell with a TextManager component were being created/reused at the same time.