Add ConfigCreatorTableCell
There is now a button to create new Configs. Adjusted removeConfig to erase the Config from the configs vector before deleting the file.
File was previously deleted first so the data could be kept in memory on the chance the file isn't properly deleted, but requesting a new Config and then clicking the delete button is inherently deleting a config that has not yet been saved and has no file. Also if there is a crash, that's probably the clearest indication that something went wrong.