Fix crash on invalid regex
Initialize regex to an empty string. Also generalize validity to `FindReplaceString::getFindRegexIsValid`.
Empty initialization would a bad regex that will crash as soon as it is used anywhere. Finally fixes the crashes while removing and creating entries, which went from inconsistent to fixed to crashing every time for some reason.